+++ CHANGELOG.md
@@ -33,7 +33,7 @@
 ## 2.2.0
 
 * `restartSubsOnConnect` method renamed to `restartSubs`.
-* Added property `clearDataOnReconnection` to `simpleDDP` class constructor.
+* Added property `clearDataOnReconnection` to `DDPClient` class constructor.
 * Docs improvments.
 
 ## 2.1.1